ST-294:Refactored viewmodel
